requirements.txt
aiohttp==3.7.4.post0
discord.py[voice]==1.7.3
PyNaCl==1.4.0

aiofiles==0.7.0
aioredis==1.3.1
ujson==4.0.2
bson==0.5.10
Pillow==8.3.1
pyppeteer==0.2.5